feat: agrega nuevos endpoints para ver datos del cliente, solicitar y reprogramar turnos

+++ src/main/java/com/oo2/grupo17/dtos/records/ContactoResponseDto.java
@@ -4,7 +4,7 @@ public record ContactoResponseDto (
 
 		Long id,
 		String email,
-		int movil,
+		long movil,
 		int telefono,
 		DireccionResponseDto direccion
